Move Saved Meets logic to separate component
